+++ src/components/company/index.ts
@@ -1,2 +1,3 @@
 export { CompanyAddForm } from './CompanyAddForm';
 export { CompanyDomainStatus } from './CompanyDomainStatus';
+export { CompanyAddView } from './CompanyAddView';
